`Integration Tests`: prevent false positives when purchasing returns 5xx
Good news: offline entitlements are amazing, and our SDK is very resilient to the backend being down.
Bad news: this is kind of a false positive, this behavior is already verified by `OfflineStoreKitIntegrationTests`.

I refactored `BaseStoreKitIntegrationTests` because it had grown a lot.
I put this verification in the `purchase` methods so all tests benefit from this.

So that `OfflineStoreKitIntegrationTests` doesn't fail, I added a parameter to ignore those there.
